requirements.txt
classy_vision==0.6.0
clip==0.2.0
continuum==1.2.1
matplotlib==3.5.2
numpy==1.19.5
pandas==1.3.5
Pillow==9.3.0 
scikit_image==0.19.2
scikit_learn
scipy==1.7.3
seaborn==0.12.1
timm==0.5.4
torch==1.11.0
torchvision==0.12.0
tqdm==4.64.0
transformers==4.19.0
vissl==0.1.6
Wand==0.6.10